Make field explorer hot reloadable
